Feat: 게시글 post, patch, 댓글 post, patch

+++ pages/articles/[id].js
@@ -2,6 +2,7 @@ import style from "@/styles/ArticleDetail.module.css";
 import { useEffect, useState } from "react";
 import { getArticle, deleteArticle } from "@/pages/api/ArticleService";
 import { useRouter } from "next/router";
+import { useAuth } from "@/contexts/AuthProvider";
 import Image from "next/image";
 import profile from "@/public/ic_profile.png";
 import heart from "@/public/ic_heart.png";
@@ -13,27 +14,40 @@ import KebabMenu from "@/components/KebabMenu";
 export default function ArticleDetail() {
   const router = useRouter();
   const { id } = router.query;
+  const { user, isPending } = useAuth(true);
   const [article, setArticle] = useState(null);
 
   const handleEditArticle = () => {
-    router.push({
-      pathname: '/post',
-      query: { id, title: article.title, content: article.content, images: article.images },
-    })
+    if (article.writer.id === user.id) {
+      router.push({
+        pathname: '/post',
+        query: { id, title: article.title, content: article.content, images: article.images },
+      });
+    } else {
+      alert("본인이 작성한 글만 수정할 수 있습니다.");
+    }
   };
 
   const handleDelete = async () => {
-    if (confirm("정말 삭제하시겠습니까?")) {
-      try {
-        await deleteArticle(id);
-        router.push('/articles');
-      } catch (error) {
-        console.error(error.message);
+    if (article.writer.id === user.id) {
+      if (confirm("정말 삭제하시겠습니까?")) {
+        try {
+          await deleteArticle(id);
+          router.push('/articles');
+        } catch (error) {
+          console.error(error.message);
+        }
       }
+    } else {
+      alert("본인이 작성한 글만 삭제할 수 있습니다.");
     }
   };
 
   useEffect(() => {
+    if (!user) {
+      return;
+    }
+
     const fetchArticle = async () => {
       if (id) {
         try {
@@ -45,9 +59,9 @@ export default function ArticleDetail() {
       }
     };
     fetchArticle();
-  }, [id]);
+  }, [id, user]);
 
-  if (!article)
+  if (!article || isPending)
     return (
       <div className={style.loading}>
         <Spinner />
@@ -64,7 +78,7 @@ export default function ArticleDetail() {
         <div className={style.dateFavSection}>
           <div className={style.profile}>
             <Image className={style.profileImg} src={profile} alt="user icon" />
-            <p className={style.userName}>총명한 판다</p>
+            <p className={style.userName}>{article.writer.nickname}</p>
             <p className={style.date}>{formatDate(article.createdAt)}</p>
           </div>
           <div className={style.favSection}>
